fix signedness issue when reading rndseed
C's integer promotion rules mean that when an unsigned char is used in
an expression, it is promoted to a signed int. So the expression
(unsigned char)c<<24 can be negative. The fix is to cast characters to
unsigned long before shifting them.

+++ solution.c
@@ -278,9 +278,10 @@ int expandsolution(solutioninfo *solution, gamesetup const *game)
     solution->flags = game->solutiondata[6];
     solution->rndslidedir = indextodir(game->solutiondata[7] & 7);
     solution->stepping = (game->solutiondata[7] >> 3) & 7;
-    solution->rndseed = game->solutiondata[8] | (game->solutiondata[9] << 8)
-					      | (game->solutiondata[10] << 16)
-					      | (game->solutiondata[11] << 24);
+    solution->rndseed = (unsigned long)game->solutiondata[8]
+		      | ((unsigned long)game->solutiondata[9] << 8)
+		      | ((unsigned long)game->solutiondata[10] << 16)
+		      | ((unsigned long)game->solutiondata[11] << 24);
 
     initmovelist(&solution->moves);
     act.when = -1;
